Renamed Files and divided code
Programming things you know
